new update

+++ src/routers/quran/quran.rs
@@ -7,12 +7,15 @@ use validator::Validate;
 #[derive(Deserialize, Validate)]
 pub struct QuranQuery {
     #[validate(range(min = 1, max = 114))]
-    from: i32,
+    from: u8,
 
     #[validate(range(min = 1, max = 114))]
-    to: i32,
+    to: u8,
 }
 
+/// Example
+///
+/// `/quran?from=1&to=10`
 #[get("/quran")]
 pub async fn quran(
     query: web::Query<QuranQuery>,
@@ -25,8 +28,10 @@ pub async fn quran(
     let result = web::block(move || {
         let mut conn = pool.get().unwrap();
 
+        // Get quran surahs
+        // example BETWEEN 1 to 100
         quran_text
-            .filter(surah.between(query.from, query.to))
+            .filter(surah.between(query.from as i32, query.to as i32))
             .load::<models::QuranText>(&mut conn)
             .unwrap()
     })
